modify func

+++ Libft/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: daykim <daykim@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/16 18:02:27 by daykim            #+#    #+#             */
-/*   Updated: 2022/01/26 20:04:10 by daykim           ###   ########.fr       */
+/*   Updated: 2022/01/27 22:16:33 by daykim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *str1, const char *str2, size_t n)
 {
-	int	index;
+	size_t	index;
 
 	index = 0;
 	while (*(str1 + index) && *(str2 + index) && (index < n))
